Rename task manager => task factory
